+++ plugins/dio_file_uploader/lib/src/file_handler.dart
@@ -17,6 +17,7 @@ class DioFileHandler extends SocketFileHandler<dio.Response<dynamic>> {
     super.body,
     super.fileKey,
     super.fileParser,
+    super.headersCallback,
     this.cancelToken,
   }) : _client = client;