+++ plugins/dio_file_uploader/lib/src/restorable_chunked_file_handler.dart
@@ -33,6 +33,7 @@ class DioRestorableChunkedFileHandler
     super.chunkSize,
     super.fileKey,
     super.chunkParser,
+    super.presentHeadersCallback,
     this.cancelToken,
   }) : _client = client;