+++ plugins/http_file_uploader/lib/src/file_handler.dart
@@ -20,6 +20,8 @@ class HttpFileHandler extends SocketFileHandler<http.Response> {
     super.headers,
     super.body,
     super.fileKey,
+    super.headersCallback,
+    super.fileParser,
   }) : _client = client;
 
   final http.Client _client;