+++ tools/file_uploader_socket_interfaces/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [2.0.0] - 2026-05-14
 
+### Added
+
+- `headersCallback` for SocketFileHandler
+- `presentHeadersCallback` for SocketRestorableChunkedFileHandler
+
 ### Changed
 
 - **Breaking**, updated to reflect the breaking changes introduced in `en_file_uploader`.